requirements.txt
pandas
numpy
openpyxl
jinja2
openai
python-dotenv
